Created a class and methods for using the JAudioTagger library
It is a very basic structure for adding/editing/deleting tags for files
but I did find the ability to allow for custom tags to be used and
multiple values to be applied to a single field. Tried to add enough
details to sufficiently explain the methods and their usage along with a
sample code of what a command for writing tag information to a file
would look like.