Got the directory-data.json file working better
Going to the "Edit" menu in the toolbar should give you an option to
edit tags on a file.

The system will save all tag/directory data to directory-data.json when
the system shuts down and will load and create objects from the data in
directory-data.json upon booting up.

In directory-data.json the "key" is the path to the file and the
"values" include a "tags" Array and a "file-type" String.

Still need to fix a bug where it applies EVERY tag to every file when
booting the program. It should instead only apply the tags relevant to
each file.