Small bug fix
The Edit Tags button will now get the name of the file of whichever row
is selected in the JTable. Before it would only get the name of the file
if it was specifically selected but now it can get the file name just by
clicking on any cell within that row.